requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.3
pydantic-settings==2.9.1
anthropic==0.52.0
python-multipart==0.0.20
pdfplumber==0.11.6
python-docx==1.1.2
fpdf2==2.8.3
python-dotenv==1.1.0
